Reference issue #31 for discerning css selector from xpath expression
Xpath will, in our case, always starts with at least a / (maybe two).
Relative xpath expression starting with a dot as no sense because there is no current node.
So search will always be for the entire document (in the steps I mean)

Css selectors will also be document-wide.
It can starts with a dot but that simply mean it's a css class we're looking for.